Moved recipes/__init__.py to recipes/core.py
